create a workflow that builds the artifact
I am using `act` to test the development of the workflow. It reports that it does not recognize the make action. This seems a difference between act and the environments that GitHub provides

Because of this I am pushing this to the repository, in order to see if the same behavior occurs